fix(audio): harden DSP engine, playback stability, and service reliability
This commit implements a series of critical fixes and hardening measures across the audio engine, playback management, and core services to improve stability and user experience.

Key changes include:

- **DSP & Audio Engine**:
  - Refactored `AudioDspEngine` to use transactional parameter mutation via `updateParams` for safer atomic updates.
  - Improved `ConvolutionReverb` with smooth cosine tapering at truncation boundaries and better sample rate resynchronization.
  - Fixed `DsdDecoder` by switching to signed 64-bit integers for CIC stages to prevent overflow/drift.
  - Optimized `LoudnessContour` with time-based exponential smoothing and reduced biquad coefficient recalculation frequency.
  - Corrected `ParametricEQ` stability by resetting filter states on bypass/unbypass transitions.
  - Fixed `Crossfeed` volume jumps by applying a makeup gain factor.

- **Playback & UI**:
  - Fixed `MiniPlayer` progress bar stutter by switching from throttled Cubit state to the unthrottled `rawPositionStream`.
  - Improved `CrossfadeManager` with a continuous exponential curve to eliminate discontinuities at $f=0$.
  - Enhanced `SleepTimerManager` with persistence to restore timers across app restarts and fixed volume ratcheting during fade-out.
  - Added debouncing to smart prefetching to prevent redundant triggers.

- **Service & Data Hardening**:
  - **YTM Service**: Implemented a 30-second post-login grace window in `YtmAccountService` to suppress false-positive authentication expiry signals during cookie propagation.
  - **Cloud Sync**: Improved sync reliability by using metadata hashes (title/artist/album/duration) instead of file paths, preventing duplicate entries on file moves.
  - **Scrobbler**: Migrated legacy `SharedPreferences` keys to a canonical format and implemented deduplication logic to prevent duplicate scrobbles during session recovery.
  - **Database**: Removed legacy FTS5 virtual tables in favor of indexed `LIKE` queries with explicit `ESCAPE` clauses for safer searching.
  - **Artwork Cache**: Fixed singleton capacity mutation in `ArtworkLruCache` to ensure consistent memory management.

+++ android/app/src/main/cpp/DsdDecoder.h
@@ -43,11 +43,11 @@ private:
     static constexpr int DECIMATION_HALF = DECIMATION_TAPS / 2;
     float decimationCoeffs_[DECIMATION_TAPS] = {};
 
-    // CIC stage 1 integrators & combs (uint32_t for well-defined mod 2^32 wrap)
+    // CIC stage 1 integrators & combs (signed 64-bit to prevent overflow/drift)
     struct CicState {
-        uint32_t int1 = 0, int2 = 0, int3 = 0;
-        uint32_t comb1 = 0, comb2 = 0, comb3 = 0;
-        uint32_t comb1_d = 0, comb2_d = 0, comb3_d = 0;
+        int64_t int1 = 0, int2 = 0, int3 = 0;
+        int64_t comb1 = 0, comb2 = 0, comb3 = 0;
+        int64_t comb1_d = 0, comb2_d = 0, comb3_d = 0;
     };
     CicState cicL_;
     CicState cicR_;
